requirements.txt
flask
flask-Cors
numpy
Pillow
tensorflow
Jinja2
pandas
Pillow

